dom/tests/mochitest/general/test_browserFrame5.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=720157
-->
<head>
  <title>Test for Bug 720157</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserFrameHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=720157">Mozilla Bug 720157</a>

<!--
  Test that data: URIs work with mozbrowserlocationchange events.
-->

<script type="application/javascript;version=1.7">
"use strict";

SimpleTest.waitForExplicitFinish();

function runTest() {
  browserFrameHelpers.setEnabledPref(true);
  browserFrameHelpers.addToWhitelist();

  var iframe = document.getElementById('iframe');
  var sawLoad = false;
  var sawLocationChange = false;

  iframe.addEventListener('mozbrowserlocationchange', function(e) {
    ok(!sawLocationChange, 'Just one locationchange event.');
    ok(!sawLoad, 'locationchange before load.');
    is(e.detail, 'data:text/html,1', "event's reported location");
    sawLocationChange = true;
  });

  iframe.addEventListener('load', function() {
    ok(sawLocationChange, 'Load after locationchange.');
    ok(!sawLoad, 'Just one load event.');
    sawLoad = true;
    SimpleTest.finish();
  });

  iframe.src = 'data:text/html,1';
}

addEventListener('load', function() { SimpleTest.executeSoon(runTest); });

</script>

<iframe id='iframe' mozbrowser></iframe>

</body>
</html>
